Robyn-MMM/mmm-platform/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
google-cloud-bigquery==3.14.1
pandas==2.2.0
numpy==1.26.3
weasyprint==60.2
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.26.0
